Add New Function buttonClicked()
Why: Needed a function to pass through in the game() so that whichever
button the player chooses to click will have the appropiate value passed
through as an argument in the playRound().

How: By adding the attribute onclick and passing a function onto each
button tag. This allowed the newly added function buttonClicked() to
run and return the proper value when ever a button is clicked by the
player.